Refactor enums: move IterableAuthResponseResult, IterableEventName, and IterableActionSource to separate files and update imports accordingly
