+++ src/enums/IterableActionSource.ts
@@ -1,0 +1,10 @@
+/**
+ * Enum representing the source of IterableAction.
+ */
+export enum IterableActionSource {
+  push = 0,
+  appLink = 1,
+  inApp = 2,
+}
+
+export default IterableActionSource;
