+++ src/enums/IterableAuthResponseResult.ts
@@ -1,0 +1,6 @@
+export enum IterableAuthResponseResult {
+  SUCCESS,
+  FAILURE,
+}
+
+export default IterableAuthResponseResult;
