+++ src/enums/IterableEventName.ts
@@ -1,0 +1,11 @@
+export enum IterableEventName {
+  handleUrlCalled = 'handleUrlCalled',
+  handleCustomActionCalled = 'handleCustomActionCalled',
+  handleInAppCalled = 'handleInAppCalled',
+  handleAuthCalled = 'handleAuthCalled',
+  receivedIterableInboxChanged = 'receivedIterableInboxChanged',
+  handleAuthSuccessCalled = 'handleAuthSuccessCalled',
+  handleAuthFailureCalled = 'handleAuthFailureCalled',
+}
+
+export default IterableEventName;
